Move Secondary Panes into a separate directory
