+++ src/components/SecondaryPanes/Expressions.js
@@ -2,10 +2,10 @@ const React = require("react");
 const { connect } = require("react-redux");
 const { bindActionCreators } = require("redux");
 const ImPropTypes = require("react-immutable-proptypes");
-const actions = require("../actions");
-const { getExpressions, getLoadedObjects, getPause } = require("../selectors");
-const CloseButton = React.createFactory(require("./shared/Button/Close"));
-const ObjectInspector = React.createFactory(require("./shared/ObjectInspector"));
+const actions = require("../../actions");
+const { getExpressions, getLoadedObjects, getPause } = require("../../selectors");
+const CloseButton = React.createFactory(require("../shared/Button/Close"));
+const ObjectInspector = React.createFactory(require("../shared/ObjectInspector"));
 const { DOM: dom, PropTypes } = React;
 
 require("./Expressions.css");
