+++ src/components/SecondaryPanes/Frames.css
