+++ src/components/SecondaryPanes/Scopes.js
@@ -2,9 +2,9 @@ const React = require("react");
 const { bindActionCreators } = require("redux");
 const { connect } = require("react-redux");
 const ImPropTypes = require("react-immutable-proptypes");
-const actions = require("../actions");
-const { getSelectedFrame, getLoadedObjects, getPause } = require("../selectors");
-const ObjectInspector = React.createFactory(require("./shared/ObjectInspector"));
+const actions = require("../../actions");
+const { getSelectedFrame, getLoadedObjects, getPause } = require("../../selectors");
+const ObjectInspector = React.createFactory(require("../shared/ObjectInspector"));
 const { DOM: dom, PropTypes } = React;
 const toPairs = require("lodash/toPairs");
