+++ src/components/SecondaryPanes/WhyPaused.css
