+++ src/components/SecondaryPanes/index.js
@@ -3,15 +3,15 @@ const { DOM: dom, PropTypes, createFactory } = React;
 const { connect } = require("react-redux");
 const { bindActionCreators } = require("redux");
 const { isEnabled } = require("devtools-config");
-const Svg = require("./shared/Svg");
+const Svg = require("../shared/Svg");
 const ImPropTypes = require("react-immutable-proptypes");
 
 const { getPause, getBreakpoints,
         getBreakpointsDisabled, getBreakpointsLoading
-      } = require("../selectors");
-const { prefs } = require("../utils/prefs");
+      } = require("../../selectors");
+const { prefs } = require("../../utils/prefs");
 
-const actions = require("../actions");
+const actions = require("../../actions");
 const WhyPaused = React.createFactory(require("./WhyPaused"));
 const Breakpoints = React.createFactory(require("./Breakpoints"));
 const Expressions = React.createFactory(require("./Expressions"));
@@ -23,7 +23,7 @@ const Scopes = isEnabled("chromeScopes")
 
 const Frames = React.createFactory(require("./Frames"));
 const EventListeners = React.createFactory(require("./EventListeners"));
-const Accordion = React.createFactory(require("./shared/Accordion"));
+const Accordion = React.createFactory(require("../shared/Accordion"));
 const CommandBar = React.createFactory(require("./CommandBar"));
 require("./SecondaryPanes.css");
